% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dm.R
\name{ddm}
\alias{ddm}
\title{Calculate the likelihood of a given Dirichelet Multiomial model}
\usage{
ddm(x, phi = NULL, scale = NULL, p = NULL, log = TRUE)
}
\arguments{
\item{phi}{overdispersion paramater. Value must be in [0,1], where phi=0
is 'pure' Multinomial distribuion and increasing values of phi lead to
increasingly over-dispersed distributions.}

\item{scale}{Scale parameters, vector of matrix.}

\item{p}{matrix or vector of proportions}

\item{log}{logical if TRUE (default) return the log likelihood}
}
\value{
numeric the (log)-likelihood given by the paramaters and data
}
\description{
Calculate the likelihood of a given Dirichelet Multiomial model
}

