% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dm.R
\name{dmdm}
\alias{dmdm}
\title{Calculate the likelihood of a given mixture of Direchelet Multiomials model}
\usage{
dmdm(x, phi = NULL, scale = NULL, p = NULL, f, log = TRUE)
}
\arguments{
\item{phi}{overdispersion paramater. Value must be in [0,1], where phi=0
is 'pure' Multinomial distribuion and increasing values of phi lead to
increasingly over-dispersed distributions.}

\item{scale}{Scale parameters, vector of matrix.}

\item{p}{matrix or vector of proportions}

\item{f}{mixture proportions}

\item{log}{logical if TRUE (default) return the log likelihood}
}
\value{
ll numeric the (log)-likelihood given by the paramaters and data

w numeric matrix with the probabilities that a given observation is
        belongs to a model-component in f
}
\description{
Calculate the likelihood of a given mixture of Direchelet Multiomials model
}

