% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit.r
\name{fit_dm}
\alias{fit_dm}
\title{Fit a dirichlet-multinomial distribution}
\usage{
fit_dm(x, phi = NULL, p = NULL, phTol = 100, phAcc = 40,
  cycles = 1000, traceLevel = 0)
}
\description{
Fit a dirichlet-multinomial distribution
}

