% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dm.R
\name{rdm}
\alias{rdm}
\title{Generate random samples of Direchelet Multinomial observations}
\usage{
rdm(n, m, phi = NULL, p = NULL, scale = NULL)
}
\arguments{
\item{n}{number of observations}

\item{m}{observation sizes vector possibly of length one.}

\item{phi}{overdispersion paramater. Value must be in [0,1], where phi=0
is 'pure' Multinomial distribuion and increasing values of phi lead to
increasingly over-dispersed distributions.}

\item{p}{matrix or vector of proportions}

\item{scale}{Scale parameters, vector of matrix.}
}
\value{
A matrix of n-rows, each row representing a random draw from
}
\description{
Generate random samples of Direchelet Multinomial observations
}
\examples{
#simulate reads with 1\% sequencing error
genotypes <- matrix( 0.01/3, nrow=4, ncol=4)
diag(genotypes) <- 0.99
round(genotyes,3)
rdm(4, 40, phi=0, p=genotypes)
rdm(4, 40, phi=0.01, p=genotypes)
}

