% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rmdm}
\alias{rmdm}
\title{Generate random samples from a mixture of Direchelet Multinomials}
\usage{
rmdm(n, m, p, f, phi = NULL, scale = NULL)
}
\arguments{
\item{n}{number of observations}

\item{m}{observation sizes vector possibly of length one.}

\item{p}{matrix or vector of proportions}

\item{f}{mixture proportions}

\item{phi}{Overdispersion paramater. Value must be in [0,1], where phi=0
is 'pure' Multinomial distribuion and increasing values of phi lead to
increasingly over-dispersed distributions.}

\item{scale}{Scale parameters, vector of matrix.}
}
\value{
A matrix of n-rows, each row representing a random draw from
}
\description{
Generate random samples from a mixture of Direchelet Multinomials
}
\examples{
#simulate reads with 1\% sequencing error
#from AT-biased genome
genotypes <- matrix( 0.01/3, nrow=4, ncol=4)
diag(genotypes) <- 0.99
nfreq <- c(0.4, 0.1, 0.1, 0.4)
rmdm(10, 40, phi=0, f=nfreq p=genotypes)
rmdm(10, 40, phi=0.01,f=nfreq,  p=genotypes)
}

