% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.mdm.R
\name{anova.mdm}
\alias{anova.mdm}
\title{Analysis of Deviance, Entropy and Diversity Tables}
\usage{
\method{anova}{mdm}(object, ..., topnote = TRUE, cols = c("df", "dev",
  "pr", "ent", "div"))
}
\arguments{
\item{object, }{objects of class \code{mdm}, usually, a result of a call to
\code{\link{mdm}}.}

\item{...}{objects of class \code{mdm}, usually, a result of a call to
\code{\link{mdm}}.}

\item{topnote}{If TRUE then model descriptions appear above the anova table,
if FALSE they appear as the first column of the table}

\item{cols}{The list of colums to print out. Defaults to all columns.}
}
\value{
The analysis of deviance, entropy and diversity for a collection of
diversity models.
}
\description{
Provides an analysis of deviance, entropy and diversity for a collection of
diversity models (outputs from \code{mdm}).
}
\details{
Specifying a single object gives a sequential analysis of deviance table for
that fit. That is, the reductions in the residual sum of deviances, plus the
residual deviance The deviances are converted into entropies and
diversities, and differences in deviances are converted into differnces in
entropies and diversities.

If more than one object is specified, the table has a row for the residual
degrees of freedom and sum of deviances for each model.  For all but the
first model, the change in degrees of freedom and deviances. This only makes
statistical sense if the models are nested.  It is conventional to list the
models from smallest to largest, but this is up to the user.
}
\note{
\code{\link{mdm}} is a modifed version of \code{\link[nnet]{multinom}}
in the \code{\link{nnet}} package.
}
\examples{

library(MDM)
data(spider6)
fit0 <- mdm(y2p(spider6[,1:6])~1,data=spider6)
fit1 <- mdm(y2p(spider6[,1:6])~Water,data=spider6)
fit2 <- mdm(y2p(spider6[,1:6])~Water+Herbs,data=spider6)
fit3 <- mdm(y2p(spider6[,1:6])~Site,data=spider6)
anova(fit0,fit1,fit2,fit3)
}
\references{
De'ath, G. (2011) \emph{The Multinomial Diversity Model: Linking
Shannon Diversity To Multiple Predictors}.

Ripley, B. D. (1996) \emph{Pattern Recognition and Neural Networks}.
Cambridge.

Venables, W. N. and Ripley, B. D. (2002) \emph{Modern Applied Statistics
with S.} Fourth edition. Springer.
}
\seealso{
\code{\link[nnet]{multinom}}, \code{\link{nnet}}
}
\keyword{mdm}
