% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDM-package.R
\docType{data}
\name{spider6}
\alias{spider6}
\title{The spider data set}
\format{A data frame with 28 observations on the following 9 variables.
\describe{
  \item{Pard.lugu}{a numeric vector}
  \item{Pard.pull}{a numeric vector}
  \item{Troc.terr}{a numeric vector}
  \item{Pard.mont}{a numeric vector}
  \item{Alop.acce}{a numeric vector}
  \item{Alop.fabr}{a numeric vector}
  \item{Water}{a numeric vector}
  \item{Herbs}{a numeric vector}
  \item{Site}{a factor with 28 levels}
}}
\source{
package mvpart
}
\usage{
spider6
}
\description{
Data set on abundances of spiders and environmental predictors. This is a
subset of a larger data comprising 12 species and 6 environmental
predictors. All variables are rated on a 0-9 scale.
}
\examples{

data(spider6)
summary(spider6)
fit0 <- mdm(y2p(spider6[,1:6])~1,data=spider6)
fit1 <- mdm(y2p(spider6[,1:6])~Water+Herbs,data=spider6)
fit2 <- mdm(y2p(spider6[,1:6])~Site,data=spider6,alpha=TRUE)
anova(fit0,fit1,fit2)

}
\references{
De'ath G. (2002) \emph{Multivariate regression trees: A new
technique for modeling species-environment relationships}. Ecology, 2002,
83:1105--1117.
}
\keyword{datasets}
